Moved mathematica_parsing functionality into generateVeffModule
